fix: many bot bugs and protocol inaccuracies
