+++ manabrew-rs/crates/manabrew-protocol/src/transport/mod.rs
@@ -14,16 +14,17 @@ pub enum AgentMessage {
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, TS)]
-#[serde(rename_all = "camelCase")]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
 #[ts(export, export_to = "transport/messages.ts")]
 pub struct StateUpdate {
     pub game_view: GameViewDto,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, TS)]
-#[serde(rename_all = "camelCase")]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
 #[ts(export, export_to = "transport/messages.ts")]
 pub struct AgentPrompt {
+    pub prompt_id: u32,
     #[serde(default, skip_serializing_if = "String::is_empty")]
     pub deciding_player_id: String,
     #[serde(default, skip_serializing_if = "Option::is_none")]
